% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.KMforCSD-package}
\alias{mlr3learners.KMforCSD}
\alias{mlr3learners.KMforCSD-package}
\title{mlr3learners.KMforCSD: An MLR3 Regression Learner that Enables Cross-Validation for the Proposed Kernel Machine.}
\description{
This package implements kernel machine regression based on a modification
    of the quadratic loss, as described in \href{https://arxiv.org/abs/1505.00991}{Kernel Machines for Current Starus Data}.
    This package is a wrapper of the KMforCSD package that enables CV for hyper-parameter tuning, and is based on the MLR3 template.
}
\author{
\strong{Maintainer}: Yael Travis-Lumer \email{yael.travis@gmail.com}

}
